Lexer: Grab longest matching sequence of speecial chars
Before this change, the lexer would grab the longest sequence
of special characters, and when the sequence ended, it would look
for a match. As a result, the lexer could never match non-space-separated
special character reserved words, i.e., "a[1]:=2".

After this change, the lexer checks upon each token consumption, when in
the special character consuming mode, to see if the overall lexeme is a
match. If it is, keep consuming until we don't have a match (we want to find
the longest matching sequence). If it is not, that means one of two things:
 - We have not found a match yet, i.e., we've consumed "=", but are looking for "=="
 - We have found a match yet, and this character just broke the match, so we need to stop

The reason we keep "going" in the first case (first point in the above list) is precisely
because the language does not have a "=" reserved word, and we still need to match "=".
This change can be generalized to let us match things like "abcd", even when "a", "ab", and
"abc" are all not part of the language. However, once we consume "abcde" (not a match), we
need to stop. This would be a problem if both of the following were reserved words:
 - "abcd"
 - "abcdef"

But "a", "ab", "abc", "abcde" were not. Our solution only lets us pick up "abcd" here, and
does not let us continue consuming sequences of a failed match after we've found a valid one.
That would likely be a bad language, though I'm not sure how that would be handled.